requirements.txt
python-telegram-bot==20.7
Pillow>=9.0
aiohttp>=3.8
python-dotenv>=1.0.0
requests>=2.28.0
pytz>=2023.3